+++ src/NLog/Config/ConfigurationItemFactory.cs
@@ -189,7 +189,7 @@ namespace NLog.Config
         public IPropertyTypeConverter PropertyTypeConverter { get; set; } = new PropertyTypeConverter();
 
         /// <summary>
-        /// Perform mesage template parsing and formatting of LogEvent messages (True = Always, False = Never, Null = Auto Detect)
+        /// Perform message template parsing and formatting of LogEvent messages (True = Always, False = Never, Null = Auto Detect)
         /// </summary>
         /// <remarks>
         /// - Null (Auto Detect) : NLog-parser checks <see cref="LogEventInfo.Message"/> for positional parameters, and will then fallback to string.Format-rendering.
@@ -307,7 +307,7 @@ namespace NLog.Config
                         var parameters = CreatePreloadParameters(preloadMethod, this);
 
                         preloadMethod.Invoke(null, parameters);
-                        InternalLogger.Debug("Preload succesfully invoked for '{0}'", type.FullName);
+                        InternalLogger.Debug("Preload successfully invoked for '{0}'", type.FullName);
                     }
                     catch (Exception e)
                     {