+++ src/NLog/Internal/FileAppenders/RetryingMultiProcessFileAppender.cs
@@ -63,7 +63,8 @@ namespace NLog.Internal.FileAppenders
         /// <param name="count">The number of bytes.</param>
         public override void Write(byte[] bytes, int offset, int count)
         {
-            using (FileStream fileStream = CreateFileStream(false))
+            int overrideBufferSize = Math.Min((count / 4096 + 1) * 4096, CreateFileParameters.BufferSize);
+            using (FileStream fileStream = CreateFileStream(false, overrideBufferSize))
             {
                 fileStream.Write(bytes, offset, count);
             }
@@ -101,7 +102,7 @@ namespace NLog.Internal.FileAppenders
         }
 
         /// <summary>
-        /// Gets the length in bytes of the file associated with the appeander.
+        /// Gets the length in bytes of the file associated with the appender.
         /// </summary>
         /// <returns>A long value representing the length of the file in bytes.</returns>
         public override long? GetFileLength()